requirements.txt
python-dotenv==1.0.1 
rdflib==7.0.0
requests==2.31.0